requirements.txt
# Python dependencies for VW Crash-to-Repair Simulator

# Core BeamNG integration
beamngpy>=1.35

# Web framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Database and data processing
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Web frontend assets serving
jinja2>=3.1.0

# Logging and monitoring
structlog>=23.0.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Optional: For real VW data integration
# google-cloud-storage>=2.10.0
# psycopg2-binary>=2.9.7